Only logged in shall edit

+++ blueprint/post.py
@@ -113,6 +113,7 @@ def browse_paged(page: int):
     )
 
 @post_bp.route('/edit/<int:post_id>', methods = ['GET', 'POST'])
+@login_required
 @post_protect
 def edit_page(post_id: int):
     form = PostForm()
@@ -122,7 +123,7 @@ def edit_page(post_id: int):
         return render_template('edit.html', form = form, post = post)
     else:
         if form.validate_on_submit():
-            if form.deleted.data and current_user.is_moderator:
+            if form.deleted.data and current_user.is_authenticated and current_user.is_moderator:
                 delete_post(post)
 
                 flash(f'Permanently deleted post #{post.id}')
